+++ lib/packages/types/color3.dart
@@ -9,6 +9,7 @@ class Color3 {
 
   external static Color3 fromRGB(int r, int g, int b);
   external static Color3 fromHSV(double h, double s, double v);
+  external List<double> toHSV();
   external static Color3 get red;
   external static Color3 get green;
   external static Color3 get blue;