+++ lib/packages/types/instance.dart
@@ -1,9 +1,13 @@
+import 'package:roblox_dart/packages/types/rbx_script_signal.dart';
+
 class Instance {
   external String name;
   external String className;
   external Instance? parent;
+  external bool archivable;
 
   external Instance(String className, [bool recursive]);
+
   external void destroy();
   external Instance? findFirstChild(String name, [bool recursive]);
   external Instance waitForChild(String name, [double? timeOut]);
@@ -18,6 +22,17 @@ class Instance {
   external Instance? findFirstChildWhichIsA(String className);
   external Instance? findFirstAncestorOfClass(String className);
   external Instance? findFirstDescendant(String name);
+  external Instance? findFirstAncestorWhichIsA(String className);
+
+  external void clearAllChildren();
+  external RBXScriptSignal<Function(Instance)> getPropertyChangedSignal(
+      String property);
+
+  external RBXScriptSignal<Function(Instance)> get childAdded;
+  external RBXScriptSignal<Function(Instance)> get childRemoved;
+  external RBXScriptSignal<Function(Instance)> get descendantAdded;
+  external RBXScriptSignal<Function(Instance)> get descendantRemoving;
+  external RBXScriptSignal<Function(Instance, Instance)> get ancestryChanged;
 
   external static T of<T extends Instance>();
 }